Added a folder called web-components and moved the Timeline class into it.
